requirement.txt
faiss-cpu
sentence-transformers
langchain
pypdf
streamlit
google-generativeai